Add fmt to kecc ir structs
